+++ music_assistant/providers/sonos_s1/__init__.py
@@ -9,19 +9,22 @@ integration for Sonos.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+import asyncio
+from typing import TYPE_CHECKING, cast
 
 from music_assistant_models.config_entries import ConfigEntry, ConfigValueType
 from music_assistant_models.enums import ConfigEntryType
+from soco.discovery import scan_network
 
 from music_assistant.constants import CONF_ENTRY_MANUAL_DISCOVERY_IPS
 
 from .constants import CONF_HOUSEHOLD_ID, CONF_NETWORK_SCAN, SUPPORTED_FEATURES
-from .provider import SonosPlayerProvider, discover_household_ids
+from .provider import SonosPlayerProvider
 
 if TYPE_CHECKING:
     from music_assistant_models.config_entries import ProviderConfig
     from music_assistant_models.provider import ProviderManifest
+    from soco import SoCo
 
     from music_assistant.mass import MusicAssistant
     from music_assistant.models import ProviderInstanceType
@@ -69,3 +72,32 @@ async def get_config_entries(
             required=False,
         ),
     )
+
+
+async def discover_household_ids(mass: MusicAssistant, prefer_s1: bool = True) -> list[str]:
+    """Discover the HouseHold ID of S1 speaker(s) the network."""
+    if cache := await mass.cache.get("sonos_household_ids"):
+        return cast("list[str]", cache)
+    household_ids: list[str] = []
+
+    def get_all_sonos_ips() -> set[SoCo]:
+        """Run full network discovery and return IP's of all devices found on the network."""
+        discovered_zones: set[SoCo] | None
+        if discovered_zones := scan_network(multi_household=True):
+            return {zone.ip_address for zone in discovered_zones}
+        return set()
+
+    all_sonos_ips = await asyncio.to_thread(get_all_sonos_ips)
+    for ip_address in all_sonos_ips:
+        async with mass.http_session.get(f"http://{ip_address}:1400/status/zp") as resp:
+            if resp.status == 200:
+                data = await resp.text()
+                if prefer_s1 and "<SWGen>2</SWGen>" in data:
+                    continue
+                if "HouseholdControlID" in data:
+                    household_id = data.split("<HouseholdControlID>")[1].split(
+                        "</HouseholdControlID>"
+                    )[0]
+                    household_ids.append(household_id)
+    await mass.cache.set("sonos_household_ids", household_ids, 3600)
+    return household_ids
